1.4 fixes
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/trunk@4503 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ cagrid-1-0/caGrid/projects/gts/src/gov/nih/nci/cagrid/gts/service/SelectStatement.java
@@ -85,9 +85,9 @@ public class SelectStatement {
 			if (o instanceof String) {
 				s.setString((index), (String) o);
 			} else if (o instanceof Long) {
-				s.setLong((index), (Long) o);
+				s.setLong((index), ((Long) o).longValue());
 			} else if (o instanceof Integer) {
-				s.setInt((index), (Integer) o);
+				s.setInt((index), ((Integer) o).intValue());
 			} else {
 				throw new Exception("Unsupported type "
 						+ o.getClass().getName());
